Added autocomplete to get training data
