+++ Code/recommender-app/templates/landing_page.html
@@ -1,0 +1,92 @@
+
+<!doctype html>
+<html>
+
+<head>
+    <title>What movies do you like?</title>
+
+    <link rel="stylesheet" href="//code.jquery.com/ui/1.12.1/themes/base/jquery-ui.css">
+    <script src="https://code.jquery.com/jquery-3.5.1.min.js" crossorigin="anonymous"></script>
+    <script src="https://code.jquery.com/ui/1.12.1/jquery-ui.min.js" crossorigin="anonymous"></script>
+    <script>
+
+        $(document).ready(function(){
+
+            $(function () {
+                $("#searchBox").autocomplete({
+                    source: function (request, response) {
+                        $.ajax({
+                            type: "POST",
+                            url: "http://localhost:5000/search",
+                            dataType: "json",
+                            cache: false,
+                            data: {
+                                q: request.term
+                            },
+                            success: function (data) {
+                                //alert(data);
+                                // console.log(data);
+                                response(data);
+                            },
+                            error: function (jqXHR, textStatus, errorThrown) {
+                                console.log(textStatus + " " + errorThrown);
+                            }
+                        });
+                    },
+                    select: function(event, ui){
+                        var ulList = $('#selectedMovies');
+                        var li = $('<li/>').text(ui.item.value).appendTo(ulList);
+                        $('#searchBox').val("");
+                        return false;
+                    },
+                    minLength: 2
+                });
+            });
+
+        });
+
+        function sendMoviesForPrediction(){
+                var movie_list = []
+                $('#selectedMovies li').each(function(){
+                    movie_list.push($(this).text());
+                });
+                var data = { "movie_list": movie_list }
+                var dataJson = JSON.stringify(data);
+                console.log(dataJson);
+                $.ajax({
+                    type: "POST",
+                    url: "http://localhost:5000/predict",
+                    dataType: "json",
+                    contentType: "application/json",
+                    cache: false,
+                    data: dataJson,
+                    success: function (response) {
+                        console.log(response);
+                    },
+                    error: function (jqXHR, textStatus, errorThrown) {
+                        console.log(textStatus + " " + errorThrown);
+                    }
+                });
+                console.log(movie_list);
+            }
+
+    </script>
+</head>
+
+<body>
+    <div style="width: 600px; margin: auto;">
+        <h2>Please select movies for training</h2>
+        <p style="width: 560px; margin: auto;">
+            <label>Search Here</label>&nbsp;&nbsp;<input type="text" name="search" id="searchBox" />
+        </p>
+    </div>
+    <div style="width: 600px; margin: auto;">
+        <h2>Selected movies : </h2>
+        <ul id = "selectedMovies">
+
+        </ul>
+    </div>
+    <div style="width: 600px; margin: auto;">
+        <input type="button" name="predict" id="predict" value="Predict" onclick="sendMoviesForPrediction();">
+    </div>
+</body>
